feat(step 4-4.1): It generates strategies! Hurray! There's still quite a few things left, but the most important part left for the MVP is the implementation of what I call "The Crucible".
Which is the final series of test (basically a validation set on
Hold-Out data) to actually have a true idea of the generalization
potential of the application.

Weaknesses I already observed is the fact the mutation is not powerful
enough, we often converge to basically the same solution or same few
solutions after even 50 generations. Basically, as soon as one strategy
starts winning, crushes against the non-real strategies and then
dominates. Once the crucible is implemented, the MVP will be complete
and I'll be able to go back to worrying about those fine-tunings.