feat(step 5): Hallelujah, the step 5 is complete and with this the full MVP is built and I even got a realistic strategy out of it.
Next step, is to going again through all of the code to catch any
quirks. The next step would be to make the grammar and mapper and all
richer to be able to generate more complex strategies (which hopefully
leads to better alpha.)

Another important thing which is likely more apt for the Run phase
is creating a more appealing UI for this app, a CLI might be good,
but a web app or something might be better (nicer to look at.)
I imagine it would play well with plots.